I didn't test myself, but I checked "Effective Java":
  When you mix primitives and boxed primitives in a single operation,
  the boxed primitive is auto-unboxed.


git-svn-id: svn+ssh://src.earth.threerings.net/everything/trunk@768 a9e240f3-4568-de11-90a4-003048321218

+++ src/java/com/threerings/everything/server/ThingIndex.java
@@ -390,9 +390,7 @@ public class ThingIndex
         public ThingInfo copyWeighted (IntMap<Float> weights)
         {
             Float adjust = weights.get(categoryId);
-            return ((adjust == null) || (adjust.floatValue() == 1f))
-                ? this
-                : new ThingInfo(this, adjust);
+            return ((adjust == null) || (adjust == 1f)) ? this : new ThingInfo(this, adjust);
         }
 
         public String toString () {
